--- api-participant/src/main/java/bio/terra/pearl/api/participant/controller/EventController.java
@@ -1,31 +1,0 @@
-package bio.terra.pearl.api.participant.controller;
-
-import bio.terra.pearl.api.participant.api.EventApi;
-import bio.terra.pearl.core.service.events.MixpanelService;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-
-@Controller
-public class EventController implements EventApi {
-  private final MixpanelService mixpanelService;
-
-  public EventController(MixpanelService mixpanelService) {
-    this.mixpanelService = mixpanelService;
-  }
-
-  @Override
-  public ResponseEntity<String> trackEvent(String data) {
-    mixpanelService.logEvent(data);
-    return ResponseEntity.accepted().build();
-  }
-
-  @Override
-  public ResponseEntity<String> trackEngage(String data) {
-    return ResponseEntity.accepted().build();
-  }
-
-  @Override
-  public ResponseEntity<String> trackGroups(String data) {
-    return ResponseEntity.accepted().build();
-  }
-}
