+++ api-participant/src/main/java/bio/terra/pearl/api/participant/controller/LoggingController.java
@@ -4,6 +4,7 @@ import bio.terra.pearl.api.participant.api.LoggingApi;
 import bio.terra.pearl.core.model.log.LogEvent;
 import bio.terra.pearl.core.model.log.LogEventSource;
 import bio.terra.pearl.core.service.LoggingService;
+import bio.terra.pearl.core.service.events.MixpanelService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.http.ResponseEntity;
@@ -11,12 +12,15 @@ import org.springframework.stereotype.Controller;
 
 @Controller
 public class LoggingController implements LoggingApi {
-  private LoggingService loggingService;
-  private ObjectMapper objectMapper;
+  private final LoggingService loggingService;
+  private final ObjectMapper objectMapper;
+  private final MixpanelService mixpanelService;
 
-  public LoggingController(LoggingService loggingService, ObjectMapper objectMapper) {
+  public LoggingController(
+      LoggingService loggingService, ObjectMapper objectMapper, MixpanelService mixpanelService) {
     this.loggingService = loggingService;
     this.objectMapper = objectMapper;
+    this.mixpanelService = mixpanelService;
   }
 
   @Override
@@ -30,5 +34,23 @@ public class LoggingController implements LoggingApi {
     } catch (JsonProcessingException e) {
       return ResponseEntity.unprocessableEntity().body(e.getMessage());
     }
+  }
+
+  @Override
+  public ResponseEntity<String> trackEvent(String data) {
+    mixpanelService.logEvent(data);
+    return ResponseEntity.accepted().build();
+  }
+
+  @Override
+  // This stub method is implemented so Mixpanel calls from the frontend do not error
+  public ResponseEntity<String> trackEngage(String data) {
+    return ResponseEntity.accepted().build();
+  }
+
+  @Override
+  // This stub method is implemented so Mixpanel calls from the frontend do not error
+  public ResponseEntity<String> trackGroups(String data) {
+    return ResponseEntity.accepted().build();
   }
 }
